make sure dtor of Client is called after handlers are completed
doing socket.close() doesn't guarantee that the completition handler
is called within the strand.
as for DoWrite, since it had no way to know the lifetime of the object
sometimes it would be called after the dtor was called, resulting on
invalid reads (detected by Valgrind), now the lifetime is extended
until after the handler finishes.