Update handling for matches (when "Best of" > 1).
Before, the server would not do sidedecking in the event of a DRAW, now
it will always do sidedecking regardless of how the duel ends (assuming
there is a next duel).

Also, add handling for non-tiebreaking matches, where the maximum number
of duels per match is fixed. This is disabled at the moment as it
requires the room's host to send the correct options when creating it.