fix sqlite race condition
according to sqlite docs, you cannot make calls to the API in a
multithreaded fashion if threading mode is set to SINGLETHREAD (even if
database connection is not shared). MULTITHREAD mode addresses that.
Added a mutex to prevent DataFromCode and ExtraFromCode to access the
same database connection in unlikely scenarios.