flatten "DuelistNames" map creation to avoid heap allocations
Previously, a bunch of allocations and sorting was needed in order to build a
std::map for those. The allocation in particular (when emplacing a new
node/element to the map) was quite problematic as it could raise an exception
despite the function being marked as noexcept.

This commit optimizes that piece of critical code by considering that there will
always be enough stack space for up to 6 duelists (hardcoded limit due to
network code limitations), completely eliding any heap allocation.