fix read-after-delete when disconnecting clients
If a client was idling when the Disconnect function is called, the async tasks
would be cancelled and it would never notify the room they were to be deleted.
This could cause issues when other function that access said clients would be
called from a different thread.

For spectator clients, if connection was idling prior to calling Disconnect,
same events would occur, which would lead to ConnectionLost *not* being emitted
and spectators set holding a now deleted client.