Revert Client socket closure updates
This reverts these commit:
  * 2cc4719b4c19e336c93c1715c67867ca789853af
  * e4ecc71042f0dfd10ccb8c221f7a784f0d4cacbc
  * part of cafbc7a34031e474b35acf1cb2d8a252ed748c4a

These changes introduced a regression, a race condition between the recv and
write threads. The current method is by no means perfect -- There is no
synchronization of cancellation between the recv and the write threads and
instead, that is left to the atomic call to the dtor of Client. Nevertheless, it
is more stable than the attempted solution and will only cause issues in cases
of extreme concurrency.