improve internal representation of lobby list
Before, an unordered_map was used to store room weak_ptrs, with an id as key,
which originally was random generated, but after the rework in
d9ef2e19644a1460a80e96407830bc90478ef947 this is no longer the case, rendering
the use of said hash map to be needlessly complex. Using a deque should improve
access speed as no hashing occurs, and should remove needless deallocations when
closing rooms, since the slots are reused as soon as possible.