Native Legend and Skill handling (#16)
Before, banlist and aliases were used to handle these types of cards introduced by Duel Links and Rush Duels. With these new changes, they are now properly handled as their own card type, instead of relying on the hack-ish behavior of banlist aliases.

Properly handle the scenario where skills or legends are taken out of the main deck: For legends that would still require adding another card in the deck, but for skills, that is not required, so there could be cases where the pre side deck has 41 cards (with the extra being the skill) and a side of 5 cards, but after siding the main deck has 40 cards with a side of 6 cards. Likewise, the opposite is possible as long as the amount of skills in the main deck is no more than the ones the player started with.

Closes #13.

Co-authored-by: Dylam De La Torre <dyxel04@gmail.com>