Hide "ghost rooms" from lobby listing
This is NOT a proper fix to the problem, it's more like hiding the symptom. The
real problem is really related to the design of the server and the protocol,
both of which are not going to be changed anytime soon... So might as well not
bother users with this.

In a nutshell, somehow it is possible to have rooms that have apparently no
users on them; When a client tries to join, it hangs as it expects more messages
from the server, but they never arrive. As far as we know, the reason for these
to exist is that ungraceful TCP disconnections are not handled properly, and TCP
keepalives are not reliable enough to detect this.